op-node: gossip flood-publish, cli gossip configuration
